Added Delete/Update Post/Comment feature

+++ server/controllers/handlerFactory.js
@@ -4,8 +4,22 @@ const APIFeatures = require('./../utils/apiFeatures');
 
 exports.deleteOne = Model =>
   catchAsync(async (req, res, next) => {
-    const doc = await Model.findByIdAndDelete(req.params.id);
+    const docToBeDeleted = await Model.findById(req.params.id);
 
+    // Check if the comment belongs to user or not
+    const headerUserId = String(req.user._id);
+    const askingUserId = String(docToBeDeleted.user._id);
+
+    if (headerUserId !== askingUserId) {
+      return next(
+        new AppError(
+          'Unauthorized. You are not allowed to delete this document!',
+          401
+        )
+      );
+    }
+
+    const doc = await Model.findByIdAndDelete(req.params.id);
     if (!doc) {
       return next(new AppError('No document found with that ID', 404));
     }
@@ -18,6 +32,21 @@ exports.deleteOne = Model =>
 
 exports.updateOne = Model =>
   catchAsync(async (req, res, next) => {
+    const docToBeUpdated = await Model.findById(req.params.id);
+
+    // Check if the comment belongs to user or not
+    const headerUserId = String(req.user._id);
+    const askingUserId = String(docToBeUpdated.user._id);
+
+    if (headerUserId !== askingUserId) {
+      return next(
+        new AppError(
+          'Unauthorized. You are not allowed to update this document!',
+          401
+        )
+      );
+    }
+
     const doc = await Model.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
       runValidators: true
